Removed need for class.
